Finished multi-module project with SDk to call the rest services.
