+++ api/src/main/java/com/ami/model/MetaData.java
@@ -1,0 +1,13 @@
+package com.ami.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * @author: Amit Khandelwal
+ * Date: 3/3/17
+ */
+public class MetaData {
+	@JsonProperty("result_count")
+	private int count;
+
+}
