+++ api/src/main/resources/fixtures/Product.json
