+++ app/src/main/java/com/ami/exceptions/EntityNotPresentException.java
