+++ app/src/main/java/com/ami/filters/LoggingFilter.java
