--- src/main/java/com/ami/api/MetaData.java
@@ -1,13 +1,0 @@
-package com.ami.api;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * @author: Amit Khandelwal
- * Date: 3/3/17
- */
-public class MetaData {
-	@JsonProperty("result_count")
-	private int count;
-
-}
