--- src/test/resources/integration-test.yml
@@ -1,36 +1,0 @@
-server:
-  type: default
-  minThreads: 20
-  maxThreads: 1024
-  requestLog:
-    appenders:
-        - type: file
-          threshold: INFO
-          timeZone: IST
-          currentLogFilename: /var/log/ami/product-service/request.log
-          archivedLogFilenamePattern: /var/log/ami/product-service/request-%d{yyyy-MM-dd-hh}.log
-          archivedFileCount: 50
-        - type: console
-          timeZone: IST
-  applicationConnectors:
-    - type: http
-      port: 9090
-      acceptorThreads: 5
-  adminConnectors:
-    - type: http
-      port: 9091
-  gzip:
-    bufferSize: 8KiB
-
-logging:
-  level: INFO
-  loggers:
-    com.ami: DEBUG
-
-#database settings
-database:
-  # the name of your JDBC driver
-  driverClass: org.h2.Driver
-  user: root
-  password: rootroot
-  url: jdbc:h2:./target/restbest
